pyproject.toml
[project]

name = "stocksai"
version = "0.0.7" 
description = "Stocks with AI" 
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE.txt"}

keywords = ["stocksai", "setuptools", "development"]
authors = [
  {name = "William Judge", email = "williamjudge94@gmail.com" }
]

classifiers = [ 
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",

  # Pick your license as you wish
  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]


dependencies = [
  "yfinance", 
  "pytest",
  "numpy",
  "matplotlib",
  "plotly-express",
  "nbformat",
  "huggingface_hub",
  "tensorflow",
  "tqdm",
  "streamsync",
  "pyarrow",
]

[project.optional-dependencies]
dev = [
    "pylint ~=2.14.0",
    "toml ~=0.10.2",
    "yapf ~=0.32.0",
]
test = [
    "pytest-cov ~=3.0.0",
]

all = [
    "stocksai[dev, test]",
]

[tool.pytest.ini_options]
pythonpath = "src"